refactor(apps-cli): rename methods
